Can find calculations by mail now

+++ src/main/java/no/ntnu/edu/stud/idatt2105/williagt/calcbackend/repository/CalculationRepository.java
@@ -15,7 +15,7 @@ public class CalculationRepository {
     @Autowired
     private JdbcTemplate jdbcTemplate;
 
-    public List<Calculations> getAllACalculationsByUserId(int userId){
+    public List<Calculations> getAllACalculationsByUserId(long userId){
         return jdbcTemplate.query("SELECT * FROM calculation WHERE userId=?",
                 BeanPropertyRowMapper.newInstance(Calculations.class), userId);
     }
